Decided 32-bit integers are enough for ISBN-10

+++ src/retail/books/ISBN10.java
@@ -3,10 +3,10 @@ package retail.books;
 public class ISBN10 extends ISBN {
 
     // TODO: Write tests for this
-    static byte checkDigit(long num) {
+    static byte checkDigit(int num) {
         return 12;
     }
-    
+
     // TODO: Write tests for this
     static boolean isValid(String s) {
         return false;
